refactor(agentic): address env interface review
Split model servers and generic parsers into extensible packages while preserving their public imports. Keep MiniGrid environment code and scenario JSONL with the task, and remove the callable dataset-path extension.
