+++ lmms_eval/agentic/servers/__init__.py
@@ -1,0 +1,12 @@
+"""Built-in model servers, split by backend.
+
+The package re-exports the stable public API so existing imports such as
+``lmms_eval.agentic.servers:OpenAIModelServer`` keep working while new
+backends can live in their own modules.
+"""
+
+from lmms_eval.agentic.servers.base import ModelServer
+from lmms_eval.agentic.servers.debug import FixedActionModelServer
+from lmms_eval.agentic.servers.openai import OpenAIModelServer
+
+__all__ = ["FixedActionModelServer", "ModelServer", "OpenAIModelServer"]
